# service/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Studio 实验室管理系统 API",
        "title": "ScienceOL Studio API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "harveyque@outlook.com"
        },
        "license": {
            "name": "GNU Affero General Public License v3.0",
            "url": "http://www.gnu.org/licenses/agpl-3.0.en.html"
        },
        "version": "1.0"
    },
    "host": "localhost:48197",
    "basePath": "/api",
    "paths": {
...
        "/v1/lab/material": {
            "post": {
                "description": "在实验室中创建物料图（节点与边）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Material"
                ],
                "summary": "创建物料图",
                "parameters": [
                    {
                        "description": "物料图创建请求",
                        "name": "material",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/material.GraphNodeReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求参数错误",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Resp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "$ref": "#/definitions/code.ErrCode"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/v1/lab/material/actions": {
            "get": {
                "description": "获取设备对应的可用动作列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Material"
                ],
                "summary": "设备可用动作",
                "parameters": [
                    {
                        "type": "string",
                        "description": "实验室UUID",
                        "name": "lab_uuid",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "设备名称",
                        "name": "name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求参数错误",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Resp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "$ref": "#/definitions/code.ErrCode"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/v1/lab/material/edge": {
            "post": {
                "description": "创建两个物料节点之间的连线",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Material"
                ],
                "summary": "创建物料连线",
                "parameters": [
                    {
                        "description": "物料连线请求",
                        "name": "edges",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/material.GraphEdge"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求参数错误",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Resp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "$ref": "#/definitions/code.ErrCode"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/v1/lab/material/edge/connection": {
            "post": {
                "description": "从边缘端创建物料之间的连线",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Material"
                ],
                "summary": "边缘端创建物料连线",
                "parameters": [
                    {
                        "description": "创建连线请求",
                        "name": "edges",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/material.CreateMaterialEdgeReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求参数错误",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Resp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "$ref": "#/definitions/code.ErrCode"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/v1/lab/material/edge/create": {
            "post": {
                "description": "从边缘端创建物料（带 UUID 的节点）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Material"
                ],
                "summary": "边缘端创建物料",
                "parameters": [
                    {
                        "description": "边缘端创建物料请求",
                        "name": "material",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/material.CreateMaterialReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求参数错误",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Resp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "$ref": "#/definitions/code.ErrCode"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/v1/lab/material/edge/upsert": {
            "post": {
                "description": "边缘端批量更新或插入物料",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Material"
                ],
                "summary": "边缘端更新/插入物料",
                "parameters": [
                    {
                        "description": "边缘端更新/插入物料请求",
                        "name": "material",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/material.UpsertMaterialReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求参数错误",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Resp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "$ref": "#/definitions/code.ErrCode"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/v1/lab/material/query": {
            "get": {
                "description": "查询实验室物料",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Material"
                ],
                "summary": "查询物料",
                "parameters": [
                    {
                        "type": "string",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "name": "with_children",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求参数错误",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Resp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "$ref": "#/definitions/code.ErrCode"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/v1/lab/material/query-by-uuid": {
            "post": {
                "description": "通过 UUID 列表查询物料信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Material"
                ],
                "summary": "按 UUID 查询物料",
                "parameters": [
                    {
                        "description": "物料 UUID 列表",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/material.MaterialQueryReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求参数错误",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Resp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "$ref": "#/definitions/code.ErrCode"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/v1/lab/material/save": {
            "post": {
                "description": "保存当前实验室的物料图",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Material"
                ],
                "summary": "保存物料图",
                "parameters": [
                    {
                        "description": "保存物料图请求",
                        "name": "material",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/material.SaveGrapReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求参数错误",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Resp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "$ref": "#/definitions/code.ErrCode"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/v1/lab/material/template/detail/{template_uuid}": {
            "get": {
                "description": "获取物料模板详情",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Material"
                ],
                "summary": "获取模板详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "模板UUID",
                        "name": "template_uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求参数错误",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Resp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "$ref": "#/definitions/code.ErrCode"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "common.Resp": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/code.ErrCode"
                },
                "data": {},
                "error": {
                    "$ref": "#/definitions/common.Error"
                },
                "timestamp": {
                    "type": "integer"
                }
            }
        },
...
        "material.CreateMaterialEdgeReq": {
            "type": "object",
            "properties": {
                "edges": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/material.MaterialEdge"
                    }
                }
            }
        },
        "material.CreateMaterialReq": {
            "type": "object",
            "properties": {
                "nodes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/material.Material"
                    }
                }
            }
        },
...
        "material.Edge": {
            "type": "object",
            "properties": {
                "source": {
                    "type": "string"
                },
                "sourceHandle": {
                    "description": "FIXME: 下面两个字段命令 unilab 需要修改命名称",
                    "type": "string"
                },
                "source_uuid": {
                    "type": "string"
                },
                "target": {
                    "type": "string"
                },
                "targetHandle": {
                    "type": "string"
                },
                "target_uuid": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "material.GraphEdge": {
            "type": "object",
            "properties": {
                "edges": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/material.Edge"
                    }
                }
            }
        },
        "material.GraphNodeReq": {
            "type": "object",
            "properties": {
                "edges": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/material.Edge"
                    }
                },
                "nodes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/material.Node"
                    }
                }
            }
        },
...
        "material.MaterialQueryReq": {
            "type": "object",
            "properties": {
                "uuids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "material.Node": {
            "type": "object",
            "required": [
                "class",
                "id",
                "name",
                "type"
            ],
            "properties": {
                "children": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "class": {
                    "type": "string"
                },
                "config": {
                    "type": "object"
                },
                "data": {
                    "type": "object"
                },
                "description": {
                    "type": "string"
                },
                "extra": {
                    "type": "object"
                },
                "id": {
                    "description": "实际是数据库的 name",
                    "type": "string"
                },
                "model": {
                    "type": "object"
                },
                "name": {
                    "description": "实际是数据库的 display name",
                    "type": "string"
                },
                "parent": {
                    "type": "string"
                },
                "parent_uuid": {
                    "type": "string"
                },
                "pose": {
                    "type": "object"
                },
                "position": {
                    "$ref": "#/definitions/model.Position"
                },
                "position_3d": {
                    "$ref": "#/definitions/model.Position"
                },
                "schema": {
                    "description": "FIXME: 这块后续要优化掉，从 reg 获取",
                    "type": "object"
                },
                "type": {
                    "$ref": "#/definitions/model.DEVICETYPE"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
...
        "material.SaveGrapReq": {
            "type": "object",
            "properties": {
                "graph": {
                    "$ref": "#/definitions/material.WSGraph"
                },
                "lab_uuid": {
                    "type": "string"
                }
            }
        },
...
        "material.UpsertMaterialReq": {
            "type": "object",
            "properties": {
                "mount_uuid": {
                    "type": "string"
                },
                "nodes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/material.Material"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}